+++ starter-code/models/Comment.js
@@ -1,0 +1,13 @@
+const mongoose = require('mongoose');
+const Schema   = mongoose.Schema;
+
+const commentSchema = Schema({
+  content: String,
+  authorId: String,
+  imagePath: String,
+  imageName: String
+});
+
+const Comment = mongoose.model('Comment', commentSchema);
+
+module.exports = Comment;
