+++ starter-code/models/User.js
@@ -4,7 +4,8 @@ const Schema   = mongoose.Schema;
 const UserSchema = Schema({
   username: String,
   email:    String,
-  password: String
+  password: String,
+  picPath: String
 });
 
 const User = mongoose.model('User', UserSchema);
