+++ starter-code/package.json
@@ -19,7 +19,7 @@
     "hbs": "^4.0.1",
     "mongoose": "^4.8.2",
     "morgan": "~1.7.0",
-    "multer": "^1.3.0",
+    "multer": "^1.4.1",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "serve-favicon": "~2.3.0"
