+++ starter-code/views/authentication/signup.hbs
@@ -2,7 +2,7 @@
   <div class="row">
     {{ message }}
     <div class="col-sm-6 centered-form">
-      <form action="/signup" method="post">
+      <form action="/signup" method="post" enctype="multipart/form-data">
         <div class="form-group">
           <label for="username">Username</label>
           <input type="text" name="username" class="form-control">
@@ -14,6 +14,10 @@
         <div class="form-group">
           <label for="password">Password</label>
           <input type="password" name="password" class="form-control">
+        </div>
+        <div class="form-group">
+          <label for="file">Your profile image</label>
+          <input type="file" name="image" class="form-control">
         </div>
         <button type="submit">Signup</button>
       </form>
