+++ starter-code/views/index.hbs
@@ -1,10 +1,2 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>{{title}}</title>
-    <link rel='stylesheet' href='/stylesheets/style.css' />
-  </head>
-  <body>
-    <h1>Welcome to IronTumblr</h1>
-  </body>
-</html>
+<a href="/login">Login</a>
+<a href="/signup">Sign up</a>
